requirements.txt
tensorflow>=2.0.0
tensorflow_hub
Pillow
matplotlib
